requirements.txt
pytube
youtube-transcript-api
python-dotenv
ffmpeg-python
openai
google-api-python-client